Change decimalfield to integerfield for membership form

+++ nadine/forms.py
@@ -485,8 +485,8 @@ class SubForm(forms.Form):
     allowance = forms.IntegerField(min_value=0, required=False)
     start_date = forms.DateField(widget=forms.TextInput(attrs={'class': 'start_date'}), required=False)
     end_date = forms.DateField(widget=forms.TextInput(attrs={'class': 'end_date'}), required=False)
-    monthly_rate = forms.DecimalField(min_value=0, decimal_places=2, required=False)
-    overage_rate = forms.DecimalField(required=False, decimal_places=2, min_value=0)
+    monthly_rate = forms.IntegerField(min_value=0, required=False)
+    overage_rate = forms.IntegerField(required=False, min_value=0)
     paid_by = forms.CharField(widget=forms.TextInput(attrs={'class': 'paying_user'}), max_length=128, required=False)
 
     def save(self):
